Splits Extension to own module
