+++ pbcore/models/assets.py
@@ -7,8 +7,6 @@ from pbcore.models.base import (
     PBCoreAttributesTime,
     PBCoreAttributesAffiliation,
     PBCoreBaseModel,
-    PBCoreTextElement,
-    PBCoreBaseAttributes,
 )
 
 # PBCore Assets
@@ -183,39 +181,7 @@ class PBCoreAnnotation(PBCoreElement):
     """PBCoreAnnotation element."""
 
     annotationType: Optional[str] = None
-
-
-class ExtensionElement(PBCoreTextElement):
-    """ExtensionElement element."""
-
-
-class ExtensionValue(PBCoreTextElement):
-    """ExtensionValue element."""
-
-
-class ExtensionAuthorityUsed(PBCoreTextElement):
-    """ExtensionAuthorityUsed element."""
-
-
-class ExtensionWrap(PBCoreBaseAttributes):
-    """ExtensionWrap element."""
-
-    extensionElement: ExtensionElement
-    extensionValue: ExtensionValue
-    extensionAuthorityUsed: Optional[ExtensionAuthorityUsed] = None
-
-
-class ExtensionEmbedded(PBCoreBaseAttributes):
-    """ExtensionEmbedded element."""
-
-
-class PBCoreExtension(PBCoreBaseModel):
-    """PBCoreExtension element.
-
-    TODO: Validate that extensionWrap and extensionEmbedded are not both present."""
-
-    extensionWrap: Optional[ExtensionWrap] = None
-    extensionEmbedded: Optional[ExtensionEmbedded] = None
+    annotation: Optional[str] = None
 
 
 __all__ = [
@@ -236,5 +202,4 @@ __all__ = [
     "PBCoreRightsSummary",
     "PBCorePart",
     "PBCoreAnnotation",
-    "PBCoreExtension",
 ]
